+++ trabalho-pratico/include/airportsManager.h
@@ -11,7 +11,7 @@ typedef struct airportsManager AirportsManager;
 //cria um novo catálogo de aeroportos
 AirportsManager *createAirportsCatalog(int size);
 //atualiza o catálogo de aeroportos
-void updateAirportCatalog(int delay, char *id_origin, char *id_destination, int id_flight, AirportsManager *airportsCatalog);
+void updateAirportCatalog(int delay, char *id_origin, char *id_destination, unsigned long int id_flight, AirportsManager *airportsCatalog);
 
 void sortAirports (AirportsManager* airportsManager);
 void sortAirportFlightsByDepartureDate_airportsCatalog(char *id, AirportsManager *airportsCatalog, void (*sortFunction)(void*,void*), void *lookup);
@@ -25,7 +25,7 @@ void destroyAirportsCatalog(AirportsManager *airportsManager);
 Airport *getAirportCatalog(AirportsManager *airportsManager, char *id);
 Hashtable *getHashtableAirportsCatalog(AirportsManager *airportsManager);
 int getAirportListSize_airportsCatalog(char *id, AirportsManager *airportsCatalog);
-int getAirportPassengersYear_airportsCatalog(int year, char *id, int (*compareFunction)(int,int,void*), int equal, void *lookup, int (*getFunction)(int,void*), AirportsManager *airportsCatalog);
+int getAirportPassengersYear_airportsCatalog(int year, char *id, int (*compareFunction)(unsigned long int,unsigned long int,void*), int equal, void *lookup, int (*getFunction)(unsigned long int,void*), AirportsManager *airportsCatalog);
 char *getNextAirportId(int index, AirportsManager *airportsCatalog);
 int getNumberAirports_airportsCatalog(AirportsManager *airportsCatalog);
 int getNumberMedianAirports(AirportsManager *airportsCatalog);
